Add modal delete user, and fix design tab category and users

+++ src/app/components/user-profile/user-profile.component.ts
@@ -4,18 +4,30 @@ import { CommonModule } from "@angular/common";
 import { FormsModule } from "@angular/forms";
 import { UserService } from "../../services/users/user.service";
 import { UserFormComponent } from "../user-form/user-form.component";
+import { ModalDeleteComponent } from "../modal-delete/modal-delete.component";
+import { Delete_Modal_Type } from "../../types/Delete_Modal";
 
 @Component({
   selector: "app-user-profile",
-  imports: [CommonModule, FormsModule, UserFormComponent],
+  imports: [CommonModule, FormsModule, UserFormComponent, ModalDeleteComponent],
   templateUrl: "./user-profile.component.html",
-  styleUrl: "./user-profile.component.css"
+  styleUrl: "./user-profile.component.css",
 })
 export class UserProfileComponent implements OnInit {
   isModalOpen: boolean = false;
   editing: boolean = false;
   selectedUser!: RegisterRequest;
+  isModalDelete: boolean = false;
+  valueDeleteModal: Delete_Modal_Type = {
+    title: "Deleting User",
+    content: "Are you sure you want to delete the user ",
+    valueBtnAccepted: "Yes, delete the user",
+    valueBtnCancelled: "Cancel, keep the user",
+    entityToDelete: null,
+  };
+
   constructor(private userService: UserService) {}
+
   ngOnInit(): void {
     this.getUser();
   }
@@ -27,20 +39,22 @@ export class UserProfileComponent implements OnInit {
     this.isModalOpen = true;
   }
 
-  deleteUser(user: RegisterRequest) {
-    if (user.id)
+  deleteUser(user: unknown) {
+    if (this.isRegisterRequest(user) && user.id) {
       this.userService.deleteUsers(user.id).subscribe({
         complete: () => {
           this.getUser();
-        }
+          this.closeModalDelete();
+        },
       });
+    }
   }
 
   getUser() {
     this.userService.getUsers().subscribe({
       next: (data) => {
         this.users = data.data;
-      }
+      },
     });
   }
   closeModal(isfetch: boolean = false) {
@@ -50,5 +64,24 @@ export class UserProfileComponent implements OnInit {
   addUser() {
     this.editing = false;
     this.isModalOpen = true;
+  }
+
+  openModalDelete(user: RegisterRequest) {
+    this.valueDeleteModal.entityToDelete = user;
+    this.isModalDelete = true;
+  }
+
+  closeModalDelete() {
+    this.isModalDelete = false;
+  }
+
+  private isRegisterRequest(obj: unknown): obj is RegisterRequest {
+    return (
+      typeof obj === "object" &&
+      obj !== null &&
+      "username" in obj &&
+      "password" in obj &&
+      "role" in obj
+    );
   }
 }
